refactor: split Map god class into GeoLayerBase + five capability mixins
Decompose the 1,733-line Map into a protected base GeoLayerBase(Scene),
which owns the display-CRS state and reproject/extract plumbing, plus
five behaviour mixins under digitalearth/scene/maps/: RasterMixin,
VectorMixin, DecorationMixin, ProjectionMixin and AnimationMixin. map.py
is now a thin composition (Map multiply-inherits all five plus the base),
so the public API and behaviour are unchanged. Repoints tests that
referenced relocated private symbols / patch targets / loggers and adds
test_map_composition.py to pin the structure.

Addresses H-6.
