+++ src/digitalearth/scene/map.py
@@ -8,66 +8,17 @@ projection is applied to the *data* upstream, not to the axes (see plan §2.4).
 The field methods here (``imshow`` and the private ``_field`` recipe) are the foundation T1.1 extends with
 ``contourf``/``contour``/``pcolormesh``/``block``.
 """
-import logging
-from typing import Any, List, Optional, Sequence, Tuple
+from digitalearth.scene.maps.animation import AnimationMixin
+from digitalearth.scene.maps.base import GeoLayerBase
+from digitalearth.scene.maps.decoration import DecorationMixin
+from digitalearth.scene.maps.projection import ProjectionMixin
+from digitalearth.scene.maps.raster import RasterMixin
+from digitalearth.scene.maps.vector import VectorMixin
 
-import numpy as np
-from shapely import MultiPoint, box, voronoi_polygons
-from shapely.affinity import scale as affine_scale
-from matplotlib.animation import FuncAnimation
-from matplotlib.cm import ScalarMappable
-from matplotlib.collections import PolyCollection
-from matplotlib.colors import Normalize
-from matplotlib.path import Path as MplPath
-from cleopatra.array_glyph import ArrayGlyph
-from cleopatra.flow_glyph import FlowGlyph
-from cleopatra.kde_glyph import KDEGlyph
-from cleopatra.mesh_glyph import MeshGlyph
-from cleopatra.polygon_glyph import PolygonGlyph
-from cleopatra.scatter_glyph import ScatterGlyph
-from cleopatra.vector_glyph import VectorGlyph
-from cleopatra.projection import apply_projection_frame
-from cleopatra.tiles import add_tiles
-from pyramids.base.crs import reproject_coordinates
-from pyramids.basemap import natural_earth
-from pyramids.dataset import Dataset
-
-from digitalearth._arrays import NAN_REDUCERS, finite, read_masked_band
-from digitalearth._crs import source_epsg
-from digitalearth.autostyle import auto_style
-from digitalearth.preprocess import add_cyclic_column
-from digitalearth.scene import projections
-from digitalearth.scene.domains import DomainLike, resolve_domain
-from digitalearth.scene.scene import Scene
-from digitalearth.sources import get_source
-from digitalearth.sources.source import Source
-
-logger = logging.getLogger(__name__)
-
-#: Cap on how many stack frames are scanned to derive a shared animation colour scale (L2).
-_CLIM_SCAN_CAP = 24
-
-#: Field-render methods accepted as the ``kind`` of an animation frame (validated up front, N1).
-_ANIMATION_KINDS = ("imshow", "contourf", "contour", "pcolormesh", "block")
-
-#: Per-cell reducers accepted by ``Map.quadtree``'s ``agg`` — the shared NaN-aware registry plus a special
-#: ``"count"`` (``len`` over the per-cell index array, ignoring the column).
-_QUADTREE_AGG = {**NAN_REDUCERS, "count": len}
+__all__ = ["Map"]
 
 
-def _stretch_to_unit(stack: np.ndarray) -> np.ndarray:
-    """Per-channel 2-98 percentile contrast stretch of an ``(rows, cols, n)`` stack into ``[0, 1]``."""
-    out = np.empty(stack.shape, dtype="float64")
-    for i in range(stack.shape[2]):
-        band = stack[..., i].astype("float64")
-        lo, hi = np.nanpercentile(band, [2, 98])
-        if hi <= lo:
-            hi = lo + 1.0
-        out[..., i] = np.clip((band - lo) / (hi - lo), 0.0, 1.0)
-    return out
-
-
-class Map(Scene):
+class Map(RasterMixin, VectorMixin, DecorationMixin, ProjectionMixin, AnimationMixin, GeoLayerBase):
     """A geospatial :class:`~digitalearth.scene.scene.Scene` that reprojects to a display CRS.
 
     Args:
@@ -109,1631 +60,3 @@ class Map(Scene):
             ```
     """
 
-    def __init__(
-        self,
-        crs: Any = 3857,
-        domain: Any = None,
-        ax: Any = None,
-        fig: Any = None,
-        figsize: Tuple[float, float] = (8, 8),
-        globe: bool = False,
-    ):
-        super().__init__(ax=ax, fig=fig, figsize=figsize)
-        self.crs = crs
-        self.domain = domain
-        self.globe = globe
-        self._graticule_lines: Optional[List[np.ndarray]] = None  # set by graticule()
-        self._last_vector: Optional[tuple] = None  # (glyph, artist, kind) of the most recent vector layer
-        self._animation: Optional[FuncAnimation] = None  # last animate()/rotate() result (kept alive, L3)
-        self._framed = False
-        self._frame_cache: Optional[tuple] = None  # (crs, (boundary, xlim, ylim)) memo
-
-    def _needs_reproject(self, dataset: Any) -> bool:
-        """Whether ``dataset`` must be reprojected to the display CRS.
-
-        Only an EPSG-int display CRS can be compared cheaply against ``dataset.epsg``. For a proj4/string
-        display CRS (e.g. an orthographic globe) we always reproject — and ``dataset.epsg`` is unreliable for
-        non-EPSG results anyway (pyramids returns 4326 for a no-code projection), so we never compare against
-        a proj4 CRS structurally here.
-
-        Args:
-            dataset: A pyramids ``Dataset`` whose ``.epsg`` is compared against the display CRS.
-
-        Returns:
-            ``False`` only when the display CRS is an ``int`` equal to ``dataset.epsg`` (data already in the
-            display CRS); ``True`` otherwise — i.e. for a differing EPSG code or any proj4/string CRS.
-        """
-        return not (isinstance(self.crs, int) and dataset.epsg == self.crs)
-
-    def _prepare(self, dataset: Any, band: int = 1) -> Source:
-        """Reproject ``dataset`` to the display CRS (if needed) and wrap it as a :class:`Source`."""
-        ds = dataset.to_crs(self.crs) if self._needs_reproject(dataset) else dataset
-        return get_source(ds, band=band)
-
-    def _field(
-        self,
-        dataset: Any,
-        *,
-        kind: str,
-        band: int = 1,
-        cmap: Optional[str] = None,
-        levels: Any = None,
-        add_colorbar: bool = False,
-        **opts,
-    ) -> Any:
-        """Render a raster ``dataset`` on the shared axes via ``cleopatra.ArrayGlyph`` (the canonical recipe).
-
-        Args:
-            dataset: A pyramids ``Dataset`` (reprojected to :attr:`crs` first).
-            kind: ArrayGlyph render kind (``auto``/``imshow``/``pcolormesh``/``contour``/``contourf``).
-            band: 1-based band index.
-            cmap: Optional colormap name.
-            levels: Optional discrete levels (int or sequence of edges).
-            add_colorbar: When ``False`` (default) the Scene owns the colorbar, not the glyph.
-            **opts: Extra styling kwargs; filtered to ``ArrayGlyph``'s accepted options.
-
-        Returns:
-            The glyph's mappable (also registered as a Scene layer).
-        """
-        src = self._prepare(dataset, band)
-        z_values, x_values, y_values = src.z.values, src.x.values, src.y.values
-        if opts.pop("cyclic", False):  # close the antimeridian seam for global fields (T5.2)
-            z_values, x_values = add_cyclic_column(z_values, x_values)
-        if cmap is None:
-            cmap = auto_style(src).get("cmap")  # per-variable default (T6.2)
-        if cmap is not None:
-            opts["cmap"] = cmap
-        if levels is not None:
-            opts["levels"] = levels
-        # Geo-reference the data. cleopatra honours `extent` only for imshow (bbox order
-        # [xmin, ymin, xmax, ymax]); contour/contourf/pcolormesh plot in array-index space unless
-        # given `coords=(x, y)`. The two are mutually exclusive, so pick by kind — otherwise the
-        # field lands at the wrong location/zoom.
-        if kind == "imshow":
-            placement = {"extent": self._extent_of(x_values, y_values)}
-        else:
-            placement = {"coords": (x_values, y_values)}
-        glyph = ArrayGlyph(
-            z_values,
-            exclude_value=[float("nan")],
-            ax=self.ax,
-            fig=self.fig,
-            **placement,
-            **opts,
-        )
-        return self._render_glyph(glyph, kind=kind, add_colorbar=add_colorbar)
-
-    def imshow(self, dataset: Any, **kwargs) -> Any:
-        """Render a raster as a pixel grid (``ArrayGlyph`` ``kind="imshow"``)."""
-        return self._field(dataset, kind="imshow", **kwargs)
-
-    def contourf(self, dataset: Any, **kwargs) -> Any:
-        """Render a raster as filled contours (``ArrayGlyph`` ``kind="contourf"``)."""
-        return self._field(dataset, kind="contourf", **kwargs)
-
-    def contour(self, dataset: Any, **kwargs) -> Any:
-        """Render a raster as line contours (``ArrayGlyph`` ``kind="contour"``)."""
-        return self._field(dataset, kind="contour", **kwargs)
-
-    def pcolormesh(self, dataset: Any, **kwargs) -> Any:
-        """Render a raster as a quadrilateral mesh (``ArrayGlyph`` ``kind="pcolormesh"``)."""
-        return self._field(dataset, kind="pcolormesh", **kwargs)
-
-    def block(self, dataset: Any, **kwargs) -> Any:
-        """Render a raster as a filled cell mesh — currently an alias of :meth:`pcolormesh`.
-
-        ``block`` is meant for *discrete* per-cell rectangles aligned to cell **edges**. cleopatra's
-        ``pcolormesh`` path accepts cell-**centre** coordinates only (it rejects edge arrays of length
-        ``n + 1``), so true edge-aligned blocks need an upstream cleopatra option; until that lands ``block``
-        draws the same filled mesh as :meth:`pcolormesh`. It is kept as a distinct, stable entry point so
-        callers and examples can switch to true blocks transparently once cleopatra supports them.
-        """
-        return self._field(dataset, kind="pcolormesh", **kwargs)
-
-    def _reproject(self, dataset: Any) -> Any:
-        """Reproject a pyramids ``Dataset`` to the display CRS (returns it unchanged when already there)."""
-        return dataset.to_crs(self.crs) if self._needs_reproject(dataset) else dataset
-
-    def _vector_input(self, features: Any, *, geom_types: Optional[Sequence[str]] = None,
-                      name: str = "layer", geom_label: Optional[str] = None) -> Any:
-        """Reproject a ``FeatureCollection`` to the display CRS, reject empty, and validate its geometry type.
-
-        Consolidates the preamble shared by the validating vector methods. Reprojects ``features`` to
-        :attr:`crs`, raises on an empty collection, and — when ``geom_types`` is given — requires every
-        geometry to be one of those types.
-
-        Args:
-            features: A pyramids ``FeatureCollection`` (reprojected to the display CRS).
-            geom_types: Allowed shapely ``geom_type`` names (e.g. ``("Point",)``); ``None`` skips the check.
-            name: The calling method's name, used in error messages.
-            geom_label: Human label for the allowed geometry in the error (defaults to the joined types).
-
-        Returns:
-            The reprojected GeoDataFrame.
-
-        Raises:
-            ValueError: if the collection is empty, or a geometry is not one of ``geom_types``.
-        """
-        gdf = features.to_crs(self.crs)
-        if len(gdf) == 0:
-            raise ValueError(f"{name} got an empty FeatureCollection (nothing to draw)")
-        if geom_types is not None and not gdf.geometry.geom_type.isin(list(geom_types)).all():
-            label = geom_label or " / ".join(geom_types)
-            raise ValueError(f"{name} requires a FeatureCollection of {label} geometries")
-        return gdf
-
-    def _polygon_layer(self, polygons: List[np.ndarray], values: Optional[np.ndarray] = None, **opts) -> Any:
-        """Draw polygons as a value-filled (``values`` given) or outline-only ``PolygonGlyph`` layer.
-
-        Consolidates the fill-vs-outline branch shared by :meth:`grid_cells`, :meth:`choropleth`,
-        :meth:`shapes`, :meth:`voronoi`, :meth:`cartogram` and :meth:`quadtree`. The Scene owns the
-        aggregated colorbar, so the glyph's own colorbar is suppressed by default.
-
-        Args:
-            polygons: Polygon rings as ``(N, 2)`` vertex arrays.
-            values: Optional per-polygon scalar values; when ``None`` only the outlines are drawn.
-            **opts: Styling kwargs forwarded to ``PolygonGlyph``.
-
-        Returns:
-            The ``PolyCollection`` (registered as a Scene layer).
-        """
-        opts.setdefault("add_colorbar", False)  # the Scene owns the aggregated colorbar
-        if values is not None:
-            glyph = PolygonGlyph(polygons, values=values, ax=self.ax, fig=self.fig, **opts)
-            return self._render_glyph(glyph, artist="plot")
-        glyph = PolygonGlyph(polygons, ax=self.ax, fig=self.fig, **opts)
-        return self._render_glyph(glyph, artist="plot", outline_only=True)
-
-    def scatter(self, features: Any, *, scale: Optional[str] = None, **opts) -> Any:
-        """Plot a pyramids ``FeatureCollection`` of points, coloured by its value column (``ScatterGlyph``).
-
-        Args:
-            features: A pyramids ``FeatureCollection`` (point geometries); reprojected to the display CRS.
-            scale: Optional column name whose values set the per-point marker size. Pair
-                it with ``size_legend=True`` (and optionally ``size_limits`` / ``size_scale``) to draw a size
-                legend. ``None`` (default) uses a single uniform marker size.
-            **opts: Styling kwargs forwarded to ``ScatterGlyph`` (``cmap``, ``scheme``, ``size_limits``,
-                ``size_scale``, ``size_legend``, ``size_legend_values``, …).
-
-        Returns:
-            The scatter ``PathCollection`` (registered as a Scene layer).
-        """
-        fc = features.to_crs(self.crs)
-        src = get_source(fc)
-        values = src.z.values if src.z is not None else None
-        sizes = np.asarray(fc[scale], dtype=float) if scale is not None else None
-        opts.setdefault("add_colorbar", False)  # the Scene owns the aggregated colorbar
-        glyph = ScatterGlyph(
-            src.x.values, src.y.values, values=values, sizes=sizes, ax=self.ax, fig=self.fig, **opts,
-        )
-        return self._render_glyph(glyph, artist="plot")
-
-    def grid_points(self, dataset: Any, **opts) -> Any:
-        """Plot raster cell centres as points coloured by value (pyramids ``to_xyz`` → ``ScatterGlyph``).
-
-        Args:
-            dataset: A pyramids ``Dataset`` (reprojected to the display CRS first).
-            **opts: Styling kwargs, filtered to ``ScatterGlyph``'s accepted options.
-
-        Returns:
-            The scatter ``PathCollection`` (registered as a Scene layer).
-
-        Examples:
-            - Scatter the raster's valid cell centres and count the resulting layer:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> from pyramids.dataset import Dataset
-                >>> from digitalearth.scene import Map
-                >>> ds = Dataset.read_file("examples/data/acc4000.tif")
-                >>> m = Map(crs=ds.epsg)
-                >>> _ = m.grid_points(ds)
-                >>> len(m.layers)
-                1
-
-                ```
-        """
-        xyz = self._reproject(dataset).to_xyz()
-        x = xyz.iloc[:, 0].to_numpy()
-        y = xyz.iloc[:, 1].to_numpy()
-        z = xyz.iloc[:, 2].to_numpy()
-        opts.setdefault("add_colorbar", False)  # the Scene owns the aggregated colorbar
-        glyph = ScatterGlyph(x, y, values=z, ax=self.ax, fig=self.fig, **opts)
-        return self._render_glyph(glyph, artist="plot")
-
-    def point_cloud(self, dataset: Any, **opts) -> Any:
-        """Alias of :meth:`grid_points` — scatter raster cell centres coloured by value."""
-        return self.grid_points(dataset, **opts)
-
-    def grid_cells(self, dataset: Any, band: int = 1, **opts) -> Any:
-        """Draw raster cells as value-coloured polygons (pyramids ``get_cell_polygons`` → ``PolygonGlyph``).
-
-        Args:
-            dataset: A pyramids ``Dataset`` (reprojected to the display CRS first).
-            band: 1-based band whose values colour the cells.
-            **opts: Styling kwargs, filtered to ``PolygonGlyph``'s accepted options.
-
-        Returns:
-            The ``PolyCollection`` (registered as a Scene layer).
-
-        Examples:
-            - Draw one polygon per raster cell and confirm the count equals rows*columns:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> from pyramids.dataset import Dataset
-                >>> from digitalearth.scene import Map
-                >>> ds = Dataset.read_file("examples/data/acc4000.tif")
-                >>> m = Map(crs=ds.epsg)
-                >>> pc = m.grid_cells(ds)
-                >>> len(pc.get_paths()) == ds.rows * ds.columns
-                True
-
-                ```
-        """
-        ds = self._reproject(dataset)
-        polygons = [np.asarray(g.exterior.coords) for g in ds.get_cell_polygons().geometry]
-        values = read_masked_band(ds, band).ravel()  # 1-based band, nodata -> NaN (shared helper)
-        polygons, values = self._finite_polygons(polygons, values)  # drop far-side cells on a globe
-        return self._polygon_layer(polygons, values, **opts)
-
-    @staticmethod
-    def _extent_of(x: Any, y: Any) -> List[float]:
-        """Return bbox-order ``[xmin, ymin, xmax, ymax]`` from 1-D x/y coordinate arrays (cleopatra order)."""
-        return [float(np.min(x)), float(np.min(y)), float(np.max(x)), float(np.max(y))]
-
-    def _extent(self, ds: Any) -> List[float]:
-        """Return bbox-order ``[xmin, ymin, xmax, ymax]`` of a dataset's cell-centre coords (cleopatra order)."""
-        return self._extent_of(ds.x, ds.y)
-
-    def rgb_composite(self, dataset: Any, bands: Sequence[int] = (1, 2, 3), **opts) -> Any:
-        """Render three raster bands as a true/false-colour RGB image (``ArrayGlyph`` RGB path).
-
-        Args:
-            dataset: A multiband pyramids ``Dataset`` (reprojected to the display CRS first).
-            bands: Three 1-based band indices mapped to R, G, B. Defaults to ``(1, 2, 3)``.
-            **opts: Styling kwargs, filtered to ``ArrayGlyph``'s accepted options.
-
-        Returns:
-            The image mappable (registered as a Scene layer).
-
-        Examples:
-            - Composite three bands of a multiband raster into one RGB image:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> import numpy as np
-                >>> from pyramids.dataset import Dataset
-                >>> from digitalearth.scene import Map
-                >>> ds = Dataset.read_file("examples/data/acc4000.tif")
-                >>> base = np.nan_to_num(ds.read_array(band=0).astype("float32"))
-                >>> rgb = Dataset.create_from_array(arr=np.stack([base, base, base]),
-                ...                                 geo=ds.geotransform, epsg=ds.epsg)
-                >>> m = Map(crs=rgb.epsg)
-                >>> _ = m.rgb_composite(rgb)
-                >>> len(m.ax.images)
-                1
-
-                ```
-        """
-        ds = self._reproject(dataset)
-        stack = np.dstack([ds.read_array(band=b - 1) for b in bands])  # (rows, cols, n)
-        # cleopatra's rgb= path is band-FIRST: it does array[rgb].transpose(1, 2, 0), so feed
-        # (n, rows, cols) and let it transpose back to (rows, cols, n) for imshow.
-        band_first = np.moveaxis(_stretch_to_unit(stack), -1, 0)
-        glyph = ArrayGlyph(
-            band_first, rgb=list(range(len(bands))), extent=self._extent(ds),
-            ax=self.ax, fig=self.fig, **opts,
-        )
-        return self._render_glyph(glyph)
-
-    def hsv_composite(self, dataset: Any, bands: Sequence[int] = (1, 2, 3), **opts) -> Any:
-        """Render three raster bands as an HSV composite (hue/sat/value → RGB → image).
-
-        Args:
-            dataset: A multiband pyramids ``Dataset`` (reprojected to the display CRS first).
-            bands: Three 1-based band indices mapped to H, S, V. Defaults to ``(1, 2, 3)``.
-            **opts: Styling kwargs, filtered to ``ArrayGlyph``'s accepted options.
-
-        Returns:
-            The image mappable (registered as a Scene layer).
-        """
-        from matplotlib.colors import hsv_to_rgb
-
-        ds = self._reproject(dataset)
-        stack = np.dstack([ds.read_array(band=b - 1) for b in bands])  # (rows, cols, n)
-        rgb = hsv_to_rgb(_stretch_to_unit(stack))                      # (rows, cols, 3) RGB
-        # band-FIRST for cleopatra's rgb= path (see rgb_composite); it transposes back to band-last.
-        band_first = np.moveaxis(rgb, -1, 0)
-        glyph = ArrayGlyph(
-            band_first, rgb=[0, 1, 2], extent=self._extent(ds), ax=self.ax, fig=self.fig,
-            **opts,
-        )
-        return self._render_glyph(glyph)
-
-    def spaghetti(self, collection: Any, band: int = 1, **opts) -> List[Any]:
-        """Overlay each member of a ``DatasetCollection`` as line contours on one axes (ensemble spaghetti).
-
-        Args:
-            collection: A pyramids ``DatasetCollection`` whose members share a grid.
-            band: 1-based band read from each member.
-            **opts: Styling kwargs forwarded to the per-member contour call.
-
-        Returns:
-            The list of per-member contour mappables (each also registered as a Scene layer).
-        """
-        return [
-            self._field(member, kind="contour", band=band, add_colorbar=False, **opts)
-            for member in collection.datasets
-        ]
-
-    def _vector(self, u_dataset: Any, v_dataset: Any, *, kind: str, band: int = 1, **opts) -> Any:
-        """Render a vector field from two rasters (u, v) on a shared grid via ``cleopatra.VectorGlyph``.
-
-        Args:
-            u_dataset: pyramids ``Dataset`` of the u (eastward) component.
-            v_dataset: pyramids ``Dataset`` of the v (northward) component.
-            kind: ``"quiver"``, ``"barbs"`` or ``"streamplot"``.
-            band: 1-based band index read from each dataset.
-            **opts: Styling kwargs, filtered to ``VectorGlyph``'s accepted options.
-
-        Returns:
-            The vector mappable (registered as a Scene layer).
-        """
-        su = self._prepare(u_dataset, band)
-        sv = self._prepare(v_dataset, band)
-        xs, ys = su.x.values, su.y.values
-        u, v = su.z.values, sv.z.values
-        # streamplot (and a tidy grid generally) needs strictly increasing axes; raster y runs
-        # north->south, so flip any descending axis and its data columns/rows to match.
-        if xs[0] > xs[-1]:
-            xs, u, v = xs[::-1], u[:, ::-1], v[:, ::-1]
-        if ys[0] > ys[-1]:
-            ys, u, v = ys[::-1], u[::-1, :], v[::-1, :]
-        x_grid, y_grid = np.meshgrid(xs, ys)
-        opts.setdefault("add_colorbar", False)  # the Scene owns the aggregated colorbar
-        glyph = VectorGlyph(
-            x_grid, y_grid, u, v, ax=self.ax, fig=self.fig, **opts,
-        )
-        im = self._render_glyph(glyph, artist="plot", kind=kind)
-        self._last_vector = (glyph, im, kind)  # remembered for quiverkey()
-        return im
-
-    def quiver(self, u_dataset: Any, v_dataset: Any, **kwargs) -> Any:
-        """Draw a vector field as arrows (``VectorGlyph`` ``kind="quiver"``)."""
-        return self._vector(u_dataset, v_dataset, kind="quiver", **kwargs)
-
-    def barbs(self, u_dataset: Any, v_dataset: Any, **kwargs) -> Any:
-        """Draw a vector field as wind barbs (``VectorGlyph`` ``kind="barbs"``)."""
-        return self._vector(u_dataset, v_dataset, kind="barbs", **kwargs)
-
-    def streamplot(self, u_dataset: Any, v_dataset: Any, **kwargs) -> Any:
-        """Draw a vector field as streamlines (``VectorGlyph`` ``kind="streamplot"``)."""
-        return self._vector(u_dataset, v_dataset, kind="streamplot", **kwargs)
-
-    def quiverkey(self, value: float, text: str, *, x: float = 0.9, y: float = 0.95, labelpos: str = "E",
-                  **kwargs) -> Any:
-        """Draw the labelled reference arrow for the most recent :meth:`quiver` layer.
-
-        Places a sample arrow of known magnitude with a label (via ``Axes.quiverkey`` on the stored quiver
-        artist), so readers can scale the field. Only ``quiver`` arrows carry a key; ``barbs``/``streamplot``
-        do not.
-
-        Args:
-            value: The reference magnitude the sample arrow represents (data units, e.g. ``10`` for 10 m/s).
-            text: The label drawn next to the arrow (e.g. ``"10 m/s"``).
-            x: Arrow x position in axes fraction (0–1). Default ``0.9``.
-            y: Arrow y position in axes fraction (0–1). Default ``0.95``.
-            labelpos: Side of the arrow for the label (``"N"``/``"S"``/``"E"``/``"W"``). Default ``"E"``.
-            **kwargs: Forwarded to ``Axes.quiverkey`` (e.g. ``coordinates``, ``color``, ``fontproperties``).
-
-        Returns:
-            The :class:`matplotlib.quiver.QuiverKey` added to the axes.
-
-        Raises:
-            ValueError: if no :meth:`quiver` layer has been drawn yet (``barbs``/``streamplot`` have no key).
-        """
-        if self._last_vector is None or self._last_vector[2] != "quiver":
-            raise ValueError("quiverkey() needs a prior quiver(...) layer (barbs/streamplot have no key)")
-        _, artist, _ = self._last_vector
-        return self.ax.quiverkey(artist, x, y, value, text, labelpos=labelpos, **kwargs)
-
-    def _reproject_point(self, lon: float, lat: float, crs: Any) -> Optional[Tuple[float, float]]:
-        """Reproject one ``(lon, lat)`` in ``crs`` to the display CRS; ``None`` if it lands off the globe.
-
-        A point on the far side of a clipped/globe display CRS reprojects to non-finite coordinates, which
-        :meth:`text` / :meth:`annotate` skip rather than drawing garbage.
-
-        Args:
-            lon: Longitude (x) in ``crs``.
-            lat: Latitude (y) in ``crs``.
-            crs: CRS of ``lon``/``lat``.
-
-        Returns:
-            The ``(x, y)`` in the display CRS, or ``None`` when the reprojected point is non-finite.
-        """
-        x, y = reproject_coordinates([lon], [lat], from_crs=crs, to_crs=self.crs)
-        if not (np.isfinite(x[0]) and np.isfinite(y[0])):
-            return None
-        return x[0], y[0]
-
-    def text(self, lon: float, lat: float, s: str, *, crs: Any = 4326, **kwargs) -> Any:
-        """Place a text label at a ``lon``/``lat`` location (reprojected to the display CRS).
-
-        The point is reprojected from ``crs`` (lon/lat by default) into the display CRS via pyramids, then
-        drawn with ``Axes.text``. On a globe, a point on the **far side** reprojects to non-finite coordinates
-        and is skipped (no artist, returns ``None``).
-
-        Args:
-            lon: Longitude (x) of the label, in ``crs``.
-            lat: Latitude (y) of the label, in ``crs``.
-            s: The text to draw.
-            crs: CRS of ``lon``/``lat`` (default ``4326`` = WGS84 lon/lat).
-            **kwargs: Forwarded to ``Axes.text`` (e.g. ``ha``, ``va``, ``fontsize``, ``color``).
-
-        Returns:
-            The :class:`matplotlib.text.Text`, or ``None`` if the point is off the visible globe.
-        """
-        xy = self._reproject_point(lon, lat, crs)
-        if xy is None:
-            return None
-        return self.ax.text(xy[0], xy[1], s, **kwargs)
-
-    def annotate(self, lon: float, lat: float, s: str, *, xytext: Any = None, crs: Any = 4326,
-                 **kwargs) -> Any:
-        """Annotate a ``lon``/``lat`` location (reprojected), optionally with an arrow.
-
-        Like :meth:`text` but via ``Axes.annotate``: the annotated point ``xy`` is the reprojected
-        ``lon``/``lat``; pass ``xytext`` (with ``arrowprops``) to draw an arrow from the label to the point.
-        A far-side point on a globe is skipped (returns ``None``).
-
-        Args:
-            lon: Longitude (x) of the annotated point, in ``crs``.
-            lat: Latitude (y) of the annotated point, in ``crs``.
-            s: The annotation text.
-            xytext: Optional text position (in the coordinate system given by ``textcoords``/``kwargs``); with
-                ``arrowprops`` an arrow is drawn from there to the point.
-            crs: CRS of ``lon``/``lat`` (default ``4326``).
-            **kwargs: Forwarded to ``Axes.annotate`` (e.g. ``arrowprops``, ``textcoords``, ``fontsize``).
-
-        Returns:
-            The :class:`matplotlib.text.Annotation`, or ``None`` if the point is off the visible globe.
-        """
-        xy = self._reproject_point(lon, lat, crs)
-        if xy is None:
-            return None
-        return self.ax.annotate(s, xy=xy, xytext=xytext, **kwargs)
-
-    def stock_img(self, dataset: Any = None, *, zorder: float = -3.0, cmap: str = "gist_earth",
-                  **kwargs) -> Any:
-        """Draw a background raster (a "stock image" backdrop) beneath all data layers.
-
-        Pass a pyramids ``Dataset`` (e.g. a low-res relief/imagery raster) to draw as the backdrop — it is
-        reprojected to the display CRS and drawn at a low ``zorder`` so data layers sit on top, and the current
-        view is preserved so a global backdrop never autoscales a regional view out. With ``dataset=None`` a
-        best-effort XYZ-tile basemap is used instead (network; skipped offline).
-
-        Note: pyramids/cleopatra ship no bundled relief raster yet, so the no-argument form relies on tiles;
-        a bundled low-res relief raster is tracked upstream (see ``stock_img`` task in the remaining-plots
-        plan). Until then, supply your own backdrop ``Dataset``.
-
-        Args:
-            dataset: A pyramids ``Dataset`` to use as the backdrop, or ``None`` to try a tile basemap.
-            zorder: Draw order for the backdrop (default ``-3.0``, below data/coastlines).
-            cmap: Colormap for a raster backdrop (ignored for the tile path).
-            **kwargs: Forwarded to :meth:`imshow` (raster) or :meth:`basemap` (tiles).
-
-        Returns:
-            The backdrop ``AxesImage`` (raster path), the tile artist, or ``None`` if a tile backdrop is
-            unavailable offline.
-        """
-        if dataset is None:
-            try:
-                return self.basemap(**kwargs)
-            except Exception as exc:  # tile servers unavailable — best-effort backdrop
-                logger.debug("stock_img tile basemap unavailable: %s", exc)
-                return None
-        with self._preserve_view():
-            im = self.imshow(dataset, cmap=cmap, **kwargs)
-            im.set_zorder(zorder)
-        return im
-
-    def _scattered(self, data: Any) -> tuple:
-        """Return ``(x, y, z)`` 1-D arrays for unstructured/point input (Dataset cells or a FeatureCollection)."""
-        if isinstance(data, Dataset):
-            xyz = self._reproject(data).to_xyz()
-            return (
-                xyz.iloc[:, 0].to_numpy(),
-                xyz.iloc[:, 1].to_numpy(),
-                xyz.iloc[:, 2].to_numpy(),
-            )
-        src = get_source(data.to_crs(self.crs))
-        if src.z is None:
-            raise ValueError("FeatureCollection has no numeric value column to contour")
-        return src.x.values, src.y.values, src.z.values
-
-    def _tri(self, data: Any, *, kind: str, **opts) -> Any:
-        """Triangulate scattered points and render via ``cleopatra.MeshGlyph``."""
-        from matplotlib.tri import Triangulation
-
-        x, y, z = self._scattered(data)
-        finite = np.isfinite(x) & np.isfinite(y)  # drop far-side points on a globe (Triangulation needs finite)
-        x, y, z = np.asarray(x)[finite], np.asarray(y)[finite], np.asarray(z)[finite]
-        tri = Triangulation(x, y)
-        glyph = MeshGlyph(x, y, tri.triangles, ax=self.ax, fig=self.fig)
-        # cleopatra 0.11.0 exposes the tripcolor/tricontour(f) artist on glyph.im (issue #2).
-        if kind == "tripcolor":
-            face_values = z[tri.triangles].mean(axis=1)
-            return self._render_glyph(glyph, face_values, location="face", colorbar=False, **opts)
-        return self._render_glyph(
-            glyph, z, location="node", filled=(kind == "tricontourf"), colorbar=False, **opts
-        )
-
-    def tricontourf(self, data: Any, **kwargs) -> Any:
-        """Filled contours of unstructured/point data (``MeshGlyph`` node data, ``filled=True``)."""
-        return self._tri(data, kind="tricontourf", **kwargs)
-
-    def tricontour(self, data: Any, **kwargs) -> Any:
-        """Line contours of unstructured/point data (``MeshGlyph`` node data, ``filled=False``)."""
-        return self._tri(data, kind="tricontour", **kwargs)
-
-    def tripcolor(self, data: Any, **kwargs) -> Any:
-        """Flat-shaded triangles of unstructured/point data (``MeshGlyph`` face data)."""
-        return self._tri(data, kind="tripcolor", **kwargs)
-
-    @staticmethod
-    def _polygon_vertices(geometry: Any) -> tuple:
-        """Return (vertex-arrays, repeat-counts) for a geopandas geometry series.
-
-        Polygons contribute their exterior ring; MultiPolygons contribute one ring per part (so a single
-        feature can map to several drawn polygons). The repeat count per feature lets callers expand a
-        per-feature value array to per-polygon.
-        """
-        polygons: List[np.ndarray] = []
-        repeats: List[int] = []
-        for geom in geometry:
-            parts = list(geom.geoms) if geom.geom_type == "MultiPolygon" else [geom]
-            polygons.extend(np.asarray(p.exterior.coords) for p in parts)
-            repeats.append(len(parts))
-        return polygons, repeats
-
-    @staticmethod
-    def _finite_polygons(polygons: List[np.ndarray], values: Optional[np.ndarray] = None) -> tuple:
-        """Drop polygons with any non-finite vertex (and the matching values).
-
-        On a projected/globe map the far hemisphere reprojects to non-finite coordinates; matplotlib's
-        ``PolyCollection`` would otherwise receive ``inf``/``nan`` vertices and fail or draw garbage. Keeps
-        only fully-finite rings, preserving order and the positional alignment between rings and ``values``.
-
-        Args:
-            polygons: Polygon rings as ``(N, 2)`` vertex arrays; any ring with an ``inf``/``nan`` vertex is
-                dropped.
-            values: Optional per-polygon scalar values, positionally aligned with ``polygons``. When given,
-                the same rings are dropped from it so the two stay aligned.
-
-        Returns:
-            ``(kept_polygons, kept_values)`` — the surviving rings, and either the filtered ``values`` array
-            or ``None`` when ``values`` was ``None``.
-
-        Examples:
-            - A ring with an ``inf`` vertex is dropped along with its value:
-                ```python
-                >>> import numpy as np
-                >>> from digitalearth.scene import Map
-                >>> good = np.array([[0.0, 0.0], [1.0, 0.0], [1.0, 1.0]])
-                >>> bad = np.array([[0.0, 0.0], [np.inf, 0.0], [1.0, 1.0]])
-                >>> kept, vals = Map._finite_polygons([good, bad, good], np.array([10.0, 20.0, 30.0]))
-                >>> len(kept)
-                2
-                >>> vals.tolist()
-                [10.0, 30.0]
-
-                ```
-            - Without values, only the finite rings come back and the second slot is ``None``:
-                ```python
-                >>> import numpy as np
-                >>> from digitalearth.scene import Map
-                >>> good = np.array([[0.0, 0.0], [1.0, 0.0], [1.0, 1.0]])
-                >>> bad = np.array([[np.nan, 0.0], [1.0, 0.0], [1.0, 1.0]])
-                >>> kept, vals = Map._finite_polygons([good, bad])
-                >>> len(kept), vals
-                (1, None)
-
-                ```
-        """
-        keep = [i for i, p in enumerate(polygons) if np.isfinite(p).all()]
-        kept = [polygons[i] for i in keep]
-        if values is None:
-            return kept, None
-        return kept, np.asarray(values)[keep]
-
-    def choropleth(self, features: Any, column: str, **opts) -> Any:
-        """Fill polygons coloured by a feature attribute (pyramids ``FeatureCollection`` → ``PolygonGlyph``).
-
-        Args:
-            features: A pyramids ``FeatureCollection`` of polygons (reprojected to the display CRS).
-            column: Name of the numeric column whose values colour the polygons.
-            **opts: Styling kwargs forwarded to ``PolygonGlyph``. Pass ``scheme`` (e.g. ``"quantiles"`` /
-                ``"fisher_jenks"``) + ``k`` to colour by discrete classes instead of a continuous scale.
-
-        Returns:
-            The ``PolyCollection`` (registered as a Scene layer).
-
-        Examples:
-            - Colour buffered point features by their ``fid`` column and count the drawn polygons:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> from pyramids.feature import FeatureCollection
-                >>> from digitalearth.scene import Map
-                >>> fc = FeatureCollection.read_file("tests/data/points.geojson")
-                >>> fc["geometry"] = fc.geometry.buffer(500.0)
-                >>> m = Map(crs=fc.epsg)
-                >>> pc = m.choropleth(fc, column="fid")
-                >>> len(pc.get_paths()) >= len(fc)
-                True
-
-                ```
-        """
-        gdf = features.to_crs(self.crs)
-        polygons, repeats = self._polygon_vertices(gdf.geometry)
-        values = np.repeat(gdf[column].to_numpy(), repeats)
-        polygons, values = self._finite_polygons(polygons, values)  # drop far-side polygons on a globe
-        return self._polygon_layer(polygons, values, **opts)
-
-    def shapes(self, features: Any, **opts) -> Any:
-        """Draw polygon outlines without fill (pyramids ``FeatureCollection`` → ``PolygonGlyph`` outline mode).
-
-        Args:
-            features: A pyramids ``FeatureCollection`` of polygons (reprojected to the display CRS).
-            **opts: Styling kwargs, filtered to ``PolygonGlyph``'s accepted options.
-
-        Returns:
-            The ``PolyCollection`` (registered as a Scene layer).
-        """
-        gdf = features.to_crs(self.crs)
-        polygons, _ = self._polygon_vertices(gdf.geometry)
-        polygons, _ = self._finite_polygons(polygons)  # drop far-side polygons on a globe
-        return self._polygon_layer(polygons, **opts)
-
-    def _clip_geometry(self, clip: Any) -> Any:
-        """Resolve a clip boundary to a single geometry in the display CRS, or ``None``.
-
-        Accepts a pyramids ``FeatureCollection`` / geopandas ``GeoDataFrame``/``GeoSeries`` (reprojected to the
-        display CRS and unioned) or a shapely geometry (assumed already in the display CRS). ``None`` → no clip.
-
-        Args:
-            clip: A ``FeatureCollection``/``GeoDataFrame``/``GeoSeries`` (reprojected + unioned), a shapely
-                geometry already in the display CRS, or ``None``.
-
-        Returns:
-            A single shapely geometry in the display CRS, or ``None`` when ``clip`` is ``None``.
-        """
-        if clip is None:
-            return None
-        if hasattr(clip, "to_crs"):  # FeatureCollection / GeoDataFrame / GeoSeries
-            geoms = clip.to_crs(self.crs)
-            geoms = geoms.geometry if hasattr(geoms, "geometry") else geoms
-            return geoms.union_all()
-        return clip  # shapely geometry, assumed already in the display CRS
-
-    @staticmethod
-    def _finite_point_xy(geom: Any, values: Optional[np.ndarray] = None):
-        """Return ``(xs, ys, values)`` for points with finite coordinates.
-
-        Points that reproject to non-finite coordinates (the far side of a clipped/globe display CRS) are
-        dropped — together with their matching ``values`` — so downstream tessellation / binning / KDE never
-        receives ``inf`` / ``nan``.
-
-        Args:
-            geom: A geopandas point ``GeoSeries`` (already in the display CRS).
-            values: Optional per-point array aligned with ``geom``; filtered by the same finite mask.
-
-        Returns:
-            tuple: ``(xs, ys, values)`` numpy arrays of finite points; ``values`` is ``None`` when not given.
-        """
-        xs = np.asarray(geom.x, dtype=float)
-        ys = np.asarray(geom.y, dtype=float)
-        mask = np.isfinite(xs) & np.isfinite(ys)
-        if values is not None:
-            values = np.asarray(values)[mask]
-        return xs[mask], ys[mask], values
-
-    def voronoi(
-        self, features: Any, column: Optional[str] = None, *, clip: Any = None, **opts
-    ) -> Any:
-        """Voronoi diagram of a point ``FeatureCollection`` (pyramids points → cells → ``PolygonGlyph``).
-
-        Tessellates the points into Voronoi cells (``shapely.voronoi_polygons`` with ``ordered=True``, so cell
-        *i* belongs to point *i*) and renders them. With ``column`` the cells are filled and coloured by that
-        point's value (like :meth:`choropleth`); without it only the cell outlines are drawn (like
-        :meth:`shapes`). Points that reproject to non-finite coordinates (the far side of a clipped/globe
-        display CRS), and duplicate points, produce no cell and are silently skipped.
-
-        Args:
-            features: A pyramids ``FeatureCollection`` of point geometries (reprojected to the display CRS).
-            column: Name of the numeric column whose value colours each cell, or ``None`` for outlines only.
-            clip: Optional boundary the cells are clipped to — a ``FeatureCollection``/``GeoDataFrame`` (reprojected
-                to the display CRS) or a shapely geometry already in the display CRS. ``None`` leaves shapely's
-                default bounded cells.
-            **opts: Styling kwargs forwarded to ``PolygonGlyph``. Pass ``scheme`` (e.g. ``"quantiles"`` /
-                ``"fisher_jenks"``) + ``k`` to colour cells by discrete classes instead of a continuous scale.
-
-        Returns:
-            The ``PolyCollection`` (registered as a Scene layer).
-
-        Raises:
-            ValueError: if ``features`` is not all single ``Point`` geometries.
-
-        Examples:
-            - Tessellate the point fixture and colour the cells by ``fid``:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> from pyramids.feature import FeatureCollection
-                >>> from digitalearth.scene import Map
-                >>> fc = FeatureCollection.read_file("tests/data/points.geojson")
-                >>> m = Map(crs=fc.epsg)
-                >>> pc = m.voronoi(fc, column="fid")
-                >>> len(m.layers)
-                1
-
-                ```
-        """
-        gdf = self._vector_input(features, geom_types=("Point",), name="voronoi", geom_label="point")
-        geom = gdf.geometry
-        col_vals = gdf[column].to_numpy() if column is not None else None
-        # Drop points with non-finite reprojected coords (far side of a clipped/globe CRS); ordered=True
-        # then keeps cell i aligned with input point i, so values map by position.
-        xs, ys, col_vals = self._finite_point_xy(geom, col_vals)
-        if xs.size == 0:
-            raise ValueError("voronoi: no finite points in the display CRS")
-        cells = voronoi_polygons(MultiPoint(list(zip(xs, ys))), ordered=True)
-        boundary = self._clip_geometry(clip)
-
-        polygons: List[np.ndarray] = []
-        values: Optional[list] = [] if column is not None else None
-        for i, cell in enumerate(cells.geoms):
-            if boundary is not None:
-                cell = cell.intersection(boundary)
-            if cell.is_empty:
-                continue
-            parts = cell.geoms if cell.geom_type.startswith("Multi") else [cell]
-            for part in parts:
-                if part.geom_type != "Polygon" or part.is_empty:
-                    continue
-                polygons.append(np.asarray(part.exterior.coords))
-                if values is not None:
-                    values.append(col_vals[i])
-        values_arr = np.asarray(values) if values is not None else None
-        polygons, values_arr = self._finite_polygons(polygons, values_arr)  # drop far-side cells on a globe
-        return self._polygon_layer(polygons, values_arr, **opts)
-
-    @staticmethod
-    def _scale_factors(values: np.ndarray, limits: Tuple[float, float]) -> np.ndarray:
-        """Linearly map ``values`` onto ``limits`` (a constant input maps to the midpoint factor).
-
-        Args:
-            values: Per-feature magnitudes to normalise.
-            limits: ``(min, max)`` output range the smallest / largest value map to.
-
-        Returns:
-            np.ndarray: Per-feature scale factors spanning ``limits``, the same shape as ``values``.
-        """
-        lo, hi = limits
-        vmin, vmax = np.nanmin(values), np.nanmax(values)
-        if vmax > vmin:
-            return lo + (values - vmin) * (hi - lo) / (vmax - vmin)
-        return np.full(values.shape, (lo + hi) / 2.0)
-
-    def cartogram(
-        self, features: Any, scale: str, column: Optional[str] = None, *,
-        limits: Tuple[float, float] = (0.2, 1.0), **opts,
-    ) -> Any:
-        """Cartogram: scale each polygon about its centroid by a value column (pyramids → ``PolygonGlyph``).
-
-        Each feature's geometry is affine-scaled about its own centroid by a factor derived from ``scale``
-        (linearly normalised across the layer to ``limits``), distorting area to encode magnitude. With
-        ``column`` the scaled polygons are filled and coloured by that column (like :meth:`choropleth`);
-        without it only the outlines are drawn (like :meth:`shapes`).
-
-        Args:
-            features: A pyramids ``FeatureCollection`` of polygon geometries (reprojected to the display CRS).
-            scale: Name of the numeric column whose value sets each feature's size (normalised to ``limits``).
-            column: Optional column whose value colours each scaled polygon, or ``None`` for outlines only.
-            limits: ``(min, max)`` scale factors mapped to the smallest/largest ``scale`` value.
-            **opts: Styling kwargs forwarded to ``PolygonGlyph``. Pass ``scheme`` (e.g. ``"quantiles"`` /
-                ``"fisher_jenks"``) + ``k`` to colour by discrete classes instead of a continuous scale.
-
-        Returns:
-            The ``PolyCollection`` (registered as a Scene layer).
-
-        Raises:
-            ValueError: if ``features`` contains non-polygon geometry.
-
-        Examples:
-            - Scale buffered points by ``fid`` and colour them by the same column:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> from pyramids.feature import FeatureCollection
-                >>> from digitalearth.scene import Map
-                >>> fc = FeatureCollection.read_file("tests/data/points.geojson")
-                >>> fc["geometry"] = fc.geometry.buffer(500.0)
-                >>> m = Map(crs=fc.epsg)
-                >>> pc = m.cartogram(fc, scale="fid", column="fid")
-                >>> len(m.layers)
-                1
-
-                ```
-        """
-        gdf = self._vector_input(features, geom_types=("Polygon", "MultiPolygon"), name="cartogram",
-                                 geom_label="polygon")
-        geom = gdf.geometry
-        factors = self._scale_factors(gdf[scale].to_numpy(dtype=float), limits)
-        scaled = [
-            affine_scale(g, xfact=f, yfact=f, origin="centroid")
-            for g, f in zip(geom, factors)
-        ]
-        polygons, repeats = self._polygon_vertices(scaled)
-        if column is not None:
-            values = np.repeat(gdf[column].to_numpy(), repeats)
-            polygons, values = self._finite_polygons(polygons, values)
-            return self._polygon_layer(polygons, values, **opts)
-        polygons, _ = self._finite_polygons(polygons)
-        return self._polygon_layer(polygons, **opts)
-
-    @staticmethod
-    def _quadtree_cells(
-        xs: np.ndarray, ys: np.ndarray, agg_fn: Any, nmax: int, nmin: int, max_depth: int = 20,
-    ) -> List[Tuple[float, float, float, float, float]]:
-        """Recursively split the points' bbox into quadrants until each cell holds ``<= nmax`` points.
-
-        Returns ``(xmin, ymin, xmax, ymax, value)`` per kept cell, where ``value = agg_fn(point_indices)``.
-        A cell with fewer than ``nmin`` points is dropped; splitting stops at ``max_depth`` and when a split
-        makes no progress (all points fall in one child), so coincident points cannot recurse forever.
-
-        Args:
-            xs: Finite point x-coordinates.
-            ys: Finite point y-coordinates, aligned with ``xs``.
-            agg_fn: Callable mapping an index array of the points in a cell to that cell's scalar value.
-            nmax: Maximum points in a cell before it is split.
-            nmin: Cells with fewer than this many points are dropped.
-            max_depth: Hard recursion-depth cap guarding against coincident points. Default 20.
-
-        Returns:
-            list[tuple]: ``(xmin, ymin, xmax, ymax, value)`` for each kept cell.
-        """
-        x0, x1 = float(np.min(xs)), float(np.max(xs))
-        y0, y1 = float(np.min(ys)), float(np.max(ys))
-        if x1 <= x0:
-            x1 = x0 + 1.0
-        if y1 <= y0:
-            y1 = y0 + 1.0
-        out: List[Tuple[float, float, float, float, float]] = []
-        stack = [(x0, y0, x1, y1, np.arange(len(xs)), 0)]
-        while stack:
-            xmin, ymin, xmax, ymax, idx, depth = stack.pop()
-            n = len(idx)
-            if n == 0:
-                continue
-            if n <= nmax or depth >= max_depth:
-                if n >= nmin:
-                    out.append((xmin, ymin, xmax, ymax, float(agg_fn(idx))))
-                continue
-            xmid, ymid = 0.5 * (xmin + xmax), 0.5 * (ymin + ymax)
-            cx, cy = xs[idx], ys[idx]
-            quads = [
-                (xmin, ymin, xmid, ymid, idx[(cx <= xmid) & (cy <= ymid)]),
-                (xmid, ymin, xmax, ymid, idx[(cx > xmid) & (cy <= ymid)]),
-                (xmin, ymid, xmid, ymax, idx[(cx <= xmid) & (cy > ymid)]),
-                (xmid, ymid, xmax, ymax, idx[(cx > xmid) & (cy > ymid)]),
-            ]
-            nonempty = [q for q in quads if len(q[4]) > 0]
-            if len(nonempty) == 1 and len(nonempty[0][4]) == n:  # no progress (coincident points)
-                if n >= nmin:
-                    out.append((xmin, ymin, xmax, ymax, float(agg_fn(idx))))
-                continue
-            for qx0, qy0, qx1, qy1, qidx in quads:
-                stack.append((qx0, qy0, qx1, qy1, qidx, depth + 1))
-        return out
-
-    def quadtree(
-        self, features: Any, column: Optional[str] = None, *, agg: Any = "mean",
-        nmax: int = 100, nmin: int = 0, clip: Any = None, **opts,
-    ) -> Any:
-        """Quadtree choropleth: aggregate points into adaptive cells (pyramids points → ``PolygonGlyph``).
-
-        Recursively splits the points' bounding box into quadrants until each cell holds ``<= nmax`` points,
-        then colours each cell by an aggregate of ``column`` (or the point **count** when ``column`` is
-        ``None``). The cells are always filled (a quadtree is a choropleth).
-
-        Args:
-            features: A pyramids ``FeatureCollection`` of point geometries (reprojected to the display CRS).
-            column: Numeric column aggregated per cell, or ``None`` to colour by point count (density).
-            agg: Per-cell reducer — one of ``"mean"``/``"sum"``/``"median"``/``"min"``/``"max"``/``"std"``/
-                ``"count"`` or a callable taking a 1-D array. Ignored when ``column`` is ``None`` (count).
-            nmax: Maximum points in a cell before it is split (smaller → finer grid).
-            nmin: Cells with fewer than this many points are dropped.
-            clip: Optional boundary the cells are clipped to (``FeatureCollection``/``GeoDataFrame`` reprojected,
-                or a shapely geometry in the display CRS). ``None`` keeps the full rectangular cells.
-            **opts: Styling kwargs forwarded to ``PolygonGlyph``. Pass ``scheme`` (e.g. ``"quantiles"`` /
-                ``"fisher_jenks"``) + ``k`` to colour cells by discrete classes instead of a continuous scale.
-
-        Returns:
-            The ``PolyCollection`` (registered as a Scene layer).
-
-        Raises:
-            ValueError: if ``features`` is not all single ``Point`` geometries, or ``agg`` is an unknown name.
-
-        Examples:
-            - Aggregate the point fixture into a fine density grid:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> from pyramids.feature import FeatureCollection
-                >>> from digitalearth.scene import Map
-                >>> fc = FeatureCollection.read_file("tests/data/points.geojson")
-                >>> m = Map(crs=fc.epsg)
-                >>> pc = m.quadtree(fc, nmax=1)
-                >>> len(m.layers)
-                1
-
-                ```
-        """
-        gdf = self._vector_input(features, geom_types=("Point",), name="quadtree", geom_label="point")
-        geom = gdf.geometry
-        # Drop points with non-finite reprojected coords (far side of a clipped/globe CRS) before binning.
-        col_vals_full = gdf[column].to_numpy(dtype=float) if column is not None else None
-        xs, ys, col_vals = self._finite_point_xy(geom, col_vals_full)
-        if xs.size == 0:
-            raise ValueError("quadtree: no finite points in the display CRS")
-        if column is None:
-            def agg_fn(idx):
-                return float(len(idx))
-        else:
-            if callable(agg):
-                reducer = agg
-            elif agg in _QUADTREE_AGG:
-                reducer = _QUADTREE_AGG[agg]
-            else:
-                raise ValueError(
-                    f"unknown agg {agg!r}; choose one of {sorted(_QUADTREE_AGG)} or a callable"
-                )
-
-            def agg_fn(idx):
-                return float(reducer(col_vals[idx]))
-
-        cells = self._quadtree_cells(xs, ys, agg_fn, nmax, nmin)
-        boundary = self._clip_geometry(clip)
-        polygons: List[np.ndarray] = []
-        values: List[float] = []
-        for xmin, ymin, xmax, ymax, val in cells:
-            if boundary is None:
-                polygons.append(
-                    np.array([[xmin, ymin], [xmax, ymin], [xmax, ymax], [xmin, ymax]])
-                )
-                values.append(val)
-                continue
-            inter = box(xmin, ymin, xmax, ymax).intersection(boundary)
-            if inter.is_empty:
-                continue
-            parts = inter.geoms if inter.geom_type.startswith("Multi") else [inter]
-            for part in parts:
-                if part.geom_type != "Polygon" or part.is_empty:
-                    continue
-                polygons.append(np.asarray(part.exterior.coords))
-                values.append(val)
-        values_arr = np.asarray(values, dtype=float)
-        polygons, values_arr = self._finite_polygons(polygons, values_arr)
-        return self._polygon_layer(polygons, values_arr, **opts)
-
-    @staticmethod
-    def _polygons_of(geom: Any) -> list:
-        """Return the ``Polygon`` parts of a shapely geometry (``[]`` for non-polygonal input).
-
-        Args:
-            geom: Any shapely geometry, or ``None``.
-
-        Returns:
-            list: ``[geom]`` for a Polygon, each Polygon part of a MultiPolygon / GeometryCollection, or
-            ``[]`` for ``None`` / non-polygonal geometry.
-        """
-        if geom is None:
-            return []
-        kind = geom.geom_type
-        if kind == "Polygon":
-            return [geom]
-        if kind in ("MultiPolygon", "GeometryCollection"):
-            return [g for g in geom.geoms if g.geom_type == "Polygon"]
-        return []
-
-    def _clip_path(self, clip: Any) -> Optional[MplPath]:
-        """Resolve a clip boundary to a matplotlib ``Path`` (data coords) for contour clipping, or ``None``.
-
-        Reuses :meth:`_clip_geometry` to reproject/union the boundary, then turns each polygon exterior ring
-        into a sub-path so a ``MultiPolygon`` clips correctly.
-
-        Args:
-            clip: A clip boundary accepted by :meth:`_clip_geometry`, or ``None``.
-
-        Returns:
-            matplotlib.path.Path | None: A path (in data coords) covering the boundary polygons, or ``None``
-            when there is no usable boundary.
-        """
-        polys = self._polygons_of(self._clip_geometry(clip))
-        verts: List[list] = []
-        codes: List[int] = []
-        for poly in polys:
-            ring = np.asarray(poly.exterior.coords)
-            if len(ring) < 3:
-                continue
-            verts.extend(ring.tolist())
-            codes.extend(
-                [MplPath.MOVETO] + [MplPath.LINETO] * (len(ring) - 2) + [MplPath.CLOSEPOLY]
-            )
-        if not verts:
-            return None
-        return MplPath(np.asarray(verts), codes)
-
-    def kde(self, features: Any, *, clip: Any = None, **opts) -> Any:
-        """2-D kernel-density (isochrone) plot of a point ``FeatureCollection`` (pyramids points → ``KDEGlyph``).
-
-        Estimates the point density on a grid and draws it as filled (``shade=True``) or line contours, coloured
-        through the shared scalar-mapping pipeline. The KDE is numpy-only (cleopatra ``KDEGlyph``).
-
-        Args:
-            features: A pyramids ``FeatureCollection`` of point geometries (reprojected to the display CRS).
-            clip: Optional boundary the density is clipped to (``FeatureCollection``/``GeoDataFrame`` reprojected,
-                or a shapely geometry in the display CRS). ``None`` draws the full grid.
-            **opts: Styling kwargs forwarded to ``KDEGlyph`` (``levels``, ``shade``, ``gridsize``,
-                ``bw_method``, ``cmap``, …).
-
-        Returns:
-            The contour set (``QuadContourSet``) registered as a Scene layer.
-
-        Raises:
-            ValueError: if ``features`` is not all single ``Point`` geometries.
-
-        Examples:
-            - Draw a density surface for the point fixture:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> from pyramids.feature import FeatureCollection
-                >>> from digitalearth.scene import Map
-                >>> fc = FeatureCollection.read_file("tests/data/points.geojson")
-                >>> m = Map(crs=fc.epsg)
-                >>> cs = m.kde(fc)
-                >>> len(m.layers)
-                1
-
-                ```
-        """
-        gdf = self._vector_input(features, geom_types=("Point",), name="kde", geom_label="point")
-        geom = gdf.geometry
-        # Drop points with non-finite reprojected coords (far side of a clipped/globe CRS) before the KDE.
-        xs, ys, _ = self._finite_point_xy(geom)
-        if xs.size == 0:
-            raise ValueError("kde: no finite points in the display CRS")
-        opts.setdefault("add_colorbar", False)  # the Scene owns the aggregated colorbar
-        glyph = KDEGlyph(
-            xs, ys, clip_path=self._clip_path(clip), ax=self.ax, fig=self.fig, **opts,
-        )
-        return self._render_glyph(glyph, artist="plot")
-
-    def sankey(
-        self, features: Any, column: Optional[str] = None, scale: Optional[str] = None, **opts,
-    ) -> Any:
-        """Spatial flow / Sankey map of a line ``FeatureCollection`` (pyramids lines → ``FlowGlyph``).
-
-        Draws each line as a path whose **colour** encodes ``column`` and whose **width** encodes ``scale``
-        (each optional). MultiLineStrings contribute one path per part.
-
-        Args:
-            features: A pyramids ``FeatureCollection`` of ``LineString``/``MultiLineString`` geometries
-                (reprojected to the display CRS).
-            column: Numeric column whose value colours each path, or ``None`` for a single colour.
-            scale: Numeric column whose value sets each path's line width, or ``None`` for a uniform width.
-            **opts: Styling kwargs forwarded to ``FlowGlyph`` (``width_limits``, ``width_scale``, ``cmap``,
-                ``size_legend``, …).
-
-        Returns:
-            The ``LineCollection`` (registered as a Scene layer).
-
-        Raises:
-            ValueError: if ``features`` contains non-line geometry.
-
-        Examples:
-            - Draw flow lines coloured and width-scaled by columns:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> import geopandas as gpd
-                >>> from shapely.geometry import LineString
-                >>> from pyramids.feature import FeatureCollection
-                >>> from digitalearth.scene import Map
-                >>> gdf = gpd.GeoDataFrame(
-                ...     {"flow": [1.0, 2.0]},
-                ...     geometry=[LineString([(0, 0), (1, 1)]), LineString([(0, 1), (1, 2)])],
-                ...     crs="EPSG:4326",
-                ... )
-                >>> m = Map(crs=4326)
-                >>> lc = m.sankey(FeatureCollection(gdf), column="flow", scale="flow")
-                >>> len(m.layers)
-                1
-
-                ```
-        """
-        gdf = self._vector_input(features, geom_types=("LineString", "MultiLineString"), name="sankey",
-                                 geom_label="line")
-        geom = gdf.geometry
-        paths: List[np.ndarray] = []
-        repeats: List[int] = []
-        for g in geom:
-            parts = list(g.geoms) if g.geom_type == "MultiLineString" else [g]
-            paths.extend(np.asarray(p.coords) for p in parts)
-            repeats.append(len(parts))
-        rep = np.asarray(repeats)
-        values = np.repeat(gdf[column].to_numpy(), rep) if column is not None else None
-        widths = np.repeat(gdf[scale].to_numpy(), rep) if scale is not None else None
-        opts.setdefault("add_colorbar", False)  # the Scene owns the aggregated colorbar
-        glyph = FlowGlyph(
-            paths, values=values, widths=widths, ax=self.ax, fig=self.fig, **opts,
-        )
-        return self._render_glyph(glyph, artist="plot")
-
-    def _project_line_features(self, fc: Any) -> List[np.ndarray]:
-        """Project a line FeatureCollection (lon/lat) to the display CRS, split at the projection limb.
-
-        Reprojecting a global line to a clipped projection (e.g. orthographic) sends the far side to
-        non-finite coordinates; per-line splitting at those gaps keeps the visible arcs and avoids the
-        ``NaN/Inf`` errors geopandas' ``clip``/``plot`` raise on such geometry.
-        """
-        from_crs = source_epsg(fc, 4326)
-        segments: List[np.ndarray] = []
-        for geom in fc.geometry:
-            if geom is None:
-                continue
-            parts = geom.geoms if geom.geom_type.startswith("Multi") else [geom]
-            for part in parts:
-                xy = np.asarray(part.coords, dtype=float)
-                if xy.size == 0:
-                    continue
-                x, y = reproject_coordinates(xy[:, 0].tolist(), xy[:, 1].tolist(),
-                                             from_crs=from_crs, to_crs=self.crs)
-                segments += projections._split_finite(np.asarray(x, float), np.asarray(y, float))
-        return segments
-
-    def _project_polygon_features(self, fc: Any) -> List[np.ndarray]:
-        """Project a polygon FeatureCollection (lon/lat) to the display CRS as finite, limb-clipped fill rings.
-
-        The fill analogue of :meth:`_project_line_features`: each exterior ring is densified, reprojected, and
-        re-closed against the projection boundary (``self._frame()[0]``) so a polygon straddling the limb
-        becomes one or more closed, fully-finite rings instead of injecting ``inf``/``nan`` into the fill.
-        Interior rings (holes) are dropped in v1 — see Digital-Earth#43.
-
-        Args:
-            fc: A pyramids ``FeatureCollection`` of polygons in lon/lat (or any CRS via ``fc.epsg``).
-
-        Returns:
-            A list of closed ``(N, 2)`` projected fill rings (empty when nothing is on the near side).
-        """
-        from_crs = source_epsg(fc, 4326)
-        boundary = self._frame()[0]
-        rings: List[np.ndarray] = []
-        for geom in fc.geometry:
-            if geom is None:
-                continue
-            parts = geom.geoms if geom.geom_type.startswith("Multi") else [geom]
-            for part in parts:
-                xy = np.asarray(part.exterior.coords, dtype=float)
-                if xy.size == 0:
-                    continue
-                xy = projections.densify_lonlat(xy, step_deg=1.0)
-                x, y = reproject_coordinates(xy[:, 0].tolist(), xy[:, 1].tolist(),
-                                             from_crs=from_crs, to_crs=self.crs)
-                rings += projections.close_visible_runs(np.asarray(x, float), np.asarray(y, float), boundary)
-        return rings
-
-    def _fill_globe_polygons(self, rings: List[np.ndarray], *, facecolor: Any, zorder: float) -> Any:
-        """Fill projected rings with a solid colour on a globe (map-specific overlay; clipped at frame time).
-
-        cleopatra ``PolygonGlyph`` only fills when given per-polygon *values*, so a uniform land/ocean fill is
-        drawn as a plain ``PolyCollection`` directly on the axes — exactly like the globe coastline ``ax.plot``
-        overlay. ``apply_projection_frame(clip_artists=True)`` clips it to the boundary at render time. The
-        view limits are preserved so a global fill never autoscales the axes out.
-
-        Args:
-            rings: Closed, finite projected fill rings (from :meth:`_project_polygon_features`).
-            facecolor: Solid fill colour.
-            zorder: Draw order (ocean below land below data below coastlines).
-
-        Returns:
-            The ``PolyCollection`` (registered as a Scene layer), or ``None`` when ``rings`` is empty.
-        """
-        if not rings:
-            return None
-        with self._preserve_view():
-            pc = PolyCollection(rings, facecolors=facecolor, edgecolors="none", zorder=zorder)
-            self.ax.add_collection(pc)
-        return self._add_layer(None, pc)
-
-    def _natural_earth(self, layer: str, resolution: str, defaults: dict, *, polygon: bool = False,
-                       zorder: float = 0.5, **kwargs) -> Any:
-        """Draw a Natural-Earth vector layer reprojected to the display CRS, clipped to the current view.
-
-        On a **globe** map, line layers (coastline/borders/rivers) are projected per-line and split at the
-        projection limb (the far side reprojects to non-finite coords) and drawn as plain polylines; polygon
-        layers (``polygon=True``: land/lakes) are projected and re-closed at the limb into finite rings and
-        filled via :meth:`_fill_globe_polygons`. Both are clipped to the boundary when the frame is applied.
-        On a **flat** map, the layer is reprojected with ``GeoDataFrame.plot``; since Natural Earth is global
-        and that autoscales the axes, the data's limits are preserved when a data layer is already present.
-
-        Args:
-            layer: Natural-Earth layer name (e.g. ``"coastline"``, ``"land"``).
-            resolution: Natural-Earth resolution (``"110m"``/``"50m"``/``"10m"``).
-            defaults: Base style; ``color``/``facecolor`` is the globe fill colour for polygon layers.
-            polygon: When True, treat the layer as filled polygons on a globe (else as lines).
-            zorder: Globe draw order for polygon fills.
-            **kwargs: Style overrides merged over ``defaults``.
-        """
-        fc = natural_earth(layer, resolution)
-        if self.globe:
-            style = {**defaults, **kwargs}
-            if polygon:
-                facecolor = style.get("facecolor", style.get("color", "#efefdb"))
-                return self._fill_globe_polygons(self._project_polygon_features(fc),
-                                                 facecolor=facecolor, zorder=zorder)
-            style.pop("edgecolor", None); style.pop("facecolor", None)
-            artists = [self.ax.plot(seg[:, 0], seg[:, 1], **style)[0] for seg in self._project_line_features(fc)]
-            return artists
-        with self._preserve_view():
-            artist = fc.to_crs(self.crs).plot(ax=self.ax, **{**defaults, **kwargs})
-        return artist
-
-    def coastlines(self, resolution: str = "110m", **kwargs) -> Any:
-        """Overlay Natural-Earth coastlines (``pyramids.basemap.natural_earth("coastline")``)."""
-        return self._natural_earth(
-            "coastline", resolution, {"color": "black", "linewidth": 0.5}, zorder=2.5, **kwargs
-        )
-
-    def borders(self, resolution: str = "110m", **kwargs) -> Any:
-        """Overlay Natural-Earth country borders."""
-        return self._natural_earth(
-            "borders", resolution, {"color": "gray", "linewidth": 0.4}, zorder=2.5, **kwargs
-        )
-
-    def land(self, resolution: str = "110m", **kwargs) -> Any:
-        """Fill Natural-Earth land polygons.
-
-        On a **flat** map the polygons are reprojected and filled directly. On a **globe** map they are
-        re-closed at the projection limb into finite rings and filled as a map-specific overlay (drawn below
-        data and coastlines, clipped to the boundary). Interior rings (holes) are dropped in v1 — see #43.
-        """
-        return self._natural_earth(
-            "land", resolution, {"color": "#efefdb", "edgecolor": "none"}, polygon=True, zorder=-1.5, **kwargs
-        )
-
-    def ocean(self, resolution: str = "110m", **kwargs) -> Any:
-        """Fill Natural-Earth ocean polygons.
-
-        On a **globe** map, ``ocean`` fills the whole projection disc (the boundary ring) with the ocean
-        colour and lets land overlay it — exact and far cheaper than clipping the global ocean polygon. On a
-        **flat** map, the Natural-Earth ocean polygons are reprojected and filled directly.
-        """
-        color = kwargs.pop("color", "#cfe6f5")
-        if self.globe:
-            boundary = self._frame()[0]
-            return self._fill_globe_polygons([np.asarray(boundary)], facecolor=color, zorder=-2.0)
-        return self._natural_earth("ocean", resolution, {"color": color, "edgecolor": "none"}, **kwargs)
-
-    def lakes(self, resolution: str = "110m", **kwargs) -> Any:
-        """Fill Natural-Earth lake polygons.
-
-        Like :meth:`land`, but with a water colour and drawn just above land (so lakes sit on the land) and
-        still below data and coastlines. On a globe the polygons are re-closed at the projection limb.
-        """
-        return self._natural_earth(
-            "lakes", resolution, {"color": "#cfe6f5", "edgecolor": "none"}, polygon=True, zorder=-1.4, **kwargs
-        )
-
-    def rivers(self, resolution: str = "110m", **kwargs) -> Any:
-        """Overlay Natural-Earth rivers (line centerlines), split at the projection limb on a globe."""
-        return self._natural_earth(
-            "rivers", resolution, {"color": "#5a8fcf", "linewidth": 0.4}, zorder=2.4, **kwargs
-        )
-
-    def basemap(self, source: Any = None, **kwargs) -> Any:
-        """Add an XYZ-tile basemap to the axes via ``cleopatra.tiles.add_tiles`` in the display CRS."""
-        return add_tiles(self.ax, source=source, crs=self.crs, **kwargs)
-
-    def set_extent(self, bbox: Sequence[float]) -> None:
-        """Set the axes extent.
-
-        Args:
-            bbox: ``[xmin, xmax, ymin, ymax]`` in the display CRS.
-        """
-        self.ax.set_xlim(bbox[0], bbox[1])
-        self.ax.set_ylim(bbox[2], bbox[3])
-
-    def set_domain(self, domain: Optional[DomainLike] = None) -> None:
-        """Set the axes extent from a named region or bbox, reprojected to the display CRS via pyramids.
-
-        Args:
-            domain: A registered region name (e.g. ``"Europe"``), an explicit ``(west, south, east, north)``
-                bbox in EPSG:4326, or ``None`` to fall back to the domain passed at construction. A no-op
-                when neither resolves to a domain.
-
-        Examples:
-            - In a geographic CRS the axes limits equal the named region's bounds:
-                ```python
-                >>> import matplotlib
-                >>> matplotlib.use("Agg")
-                >>> from digitalearth.scene import Map
-                >>> m = Map(crs=4326)
-                >>> m.set_domain("europe")
-                >>> [float(v) for v in m.ax.get_xlim()]
-                [-25.0, 45.0]
-                >>> [float(v) for v in m.ax.get_ylim()]
-                [34.0, 72.0]
-
-                ```
-        """
-        bbox = resolve_domain(domain if domain is not None else self.domain)
-        if bbox is None:
-            return
-        west, south, east, north = bbox
-        xs, ys = reproject_coordinates(
-            [west, east, west, east], [south, south, north, north], from_crs=4326, to_crs=self.crs
-        )
-        self.set_extent([min(xs), max(xs), min(ys), max(ys)])
-
-    # ------------------------------------------------------------------ globe / projection frame
-
-    def graticule(self, lon_step: float = 30.0, lat_step: float = 30.0) -> None:
-        """Add a lon/lat graticule to a projected map (drawn when the frame is applied).
-
-        Args:
-            lon_step: Meridian spacing in degrees.
-            lat_step: Parallel spacing in degrees.
-        """
-        self._graticule_lines = projections.graticule(self.crs, lon_step=lon_step, lat_step=lat_step)
-
-    def _frame(self) -> tuple:
-        """Return the cached ``(boundary, xlim, ylim)`` for the display CRS (computed once per CRS).
-
-        ``projection_frame`` reprojects a dense lon/lat sample of the whole sphere, so it is memoised here to
-        avoid recomputing it for both ``set_global`` and ``_apply_frame``. The cache is keyed on the display
-        CRS and recomputed only when the CRS changes.
-
-        Returns:
-            The ``(boundary_xy, (xmin, xmax), (ymin, ymax))`` tuple from
-            :func:`digitalearth.scene.projections.projection_frame` for the current display CRS — a closed
-            ``(N, 2)`` boundary ring plus the projected x/y limits.
-        """
-        if self._frame_cache is None or self._frame_cache[0] != self.crs:
-            self._frame_cache = (self.crs, projections.projection_frame(self.crs))
-        return self._frame_cache[1]
-
-    def set_global(self) -> None:
-        """Set the axes extent to the full projection domain (the whole globe/world)."""
-        _, xlim, ylim = self._frame()
-        self.set_extent([xlim[0], xlim[1], ylim[0], ylim[1]])
-
-    def _apply_frame(self) -> Any:
-        """Draw the projection boundary + graticule and clip the layers to it (once, at render time)."""
-        if not self.globe or self._framed:
-            return None
-        boundary, xlim, ylim = self._frame()
-        patch = apply_projection_frame(
-            self.ax, boundary_xy=boundary, xlim=xlim, ylim=ylim,
-            graticule_lines=self._graticule_lines,
-        )
-        self._framed = True
-        return patch
-
-    def render(self) -> None:
-        """Apply the projection frame if this is a globe map (idempotent). Call before showing/saving."""
-        self._apply_frame()
-
-    def save(self, path: str, **kwargs) -> None:
-        """Apply the projection frame (for a globe map) then save the figure."""
-        self._apply_frame()
-        super().save(path, **kwargs)
-
-    def show(self) -> None:
-        """Apply the projection frame (for a globe map) then show the figure."""
-        self._apply_frame()
-        super().show()
-
-    def _animate_frames(self, draw_one: Any, n_frames: int, fps: float) -> FuncAnimation:
-        """Drive ``n_frames`` of ``draw_one(i)`` on this Map's axes as a :class:`FuncAnimation`.
-
-        Each frame clears the axes and resets the per-frame layer/frame state, calls ``draw_one(i)`` to draw
-        frame ``i``, then (on a globe) sets the full-domain extent and applies the projection frame. No
-        colorbar is added per frame — pass a fixed ``vmin``/``vmax`` to keep colours stable instead.
-        """
-        def _f(i: int) -> None:
-            self.ax.clear()
-            self.layers = []
-            self._framed = False
-            draw_one(i)
-            if self.globe:
-                self.set_global()
-                self._apply_frame()
-
-        anim = FuncAnimation(self.fig, _f, frames=n_frames, interval=1000.0 / fps, blit=False)
-        self._animation = anim  # keep a strong reference so it isn't garbage-collected before save (L3)
-        return anim
-
-    @staticmethod
-    def _stack_clim(datasets: Sequence[Any]) -> Tuple[float, float]:
-        """Return the ``(min, max)`` of the first band across ``datasets``, ignoring nodata/non-finite."""
-        lows: List[float] = []
-        highs: List[float] = []
-        for ds in datasets:
-            arr = finite(read_masked_band(ds, band=1))
-            if arr.size:
-                lows.append(float(arr.min()))
-                highs.append(float(arr.max()))
-        return (min(lows), max(highs)) if lows else (0.0, 1.0)
-
-    def _resolve_animation_clim(self, datasets: Sequence[Any], opts: dict) -> None:
-        """Ensure ``opts`` carries a shared ``vmin``/``vmax`` so every animation frame uses one colour scale.
-
-        Without this, each frame's renderer auto-scales to its own data range, so the colours (and any
-        colorbar) flicker between frames. Any ``vmin``/``vmax`` already in ``opts`` is kept; a missing bound
-        is filled once from the stack (ignoring nodata/non-finite) and written back, so all frames — and the
-        colorbar — share it. Passing both ``vmin`` and ``vmax`` skips the scan entirely. To bound the cost on
-        large stacks, at most :data:`_CLIM_SCAN_CAP` evenly-spaced frames are scanned (L2).
-        """
-        vmin, vmax = opts.get("vmin"), opts.get("vmax")
-        if vmin is None or vmax is None:
-            seq = list(datasets)
-            stride = max(1, len(seq) // _CLIM_SCAN_CAP)  # cap the scan to ~_CLIM_SCAN_CAP frames
-            lo, hi = self._stack_clim(seq[::stride])
-            opts["vmin"] = lo if vmin is None else vmin
-            opts["vmax"] = hi if vmax is None else vmax
-
-    def _animation_colorbar(self, opts: dict, label: Optional[str]) -> Any:
-        """Add one static colorbar for an animation from the already-resolved ``cmap``/``vmin``/``vmax``.
-
-        The colorbar lives on its own figure axes (not the data axes that each frame clears), so it persists
-        across frames. Call :meth:`_resolve_animation_clim` first so ``opts`` has the shared clim.
-        """
-        cmap = opts.setdefault("cmap", "viridis")
-        mappable = ScalarMappable(norm=Normalize(vmin=opts.get("vmin"), vmax=opts.get("vmax")), cmap=cmap)
-        mappable.set_array([])
-        cbar = self.fig.colorbar(mappable, ax=self.ax)
-        if label is not None:
-            cbar.set_label(label)
-        return cbar
-
-    def _prime_animation(self, datasets: Sequence[Any], opts: dict, *, colorbar: bool,
-                         cbar_label: Optional[str]) -> None:
-        """Resolve one shared colour scale into ``opts`` and, if asked, add the single static colorbar.
-
-        The setup shared by :meth:`animate` and :meth:`rotate`: fill a missing ``vmin``/``vmax`` once from the
-        stack so colours don't flicker between frames, then optionally draw one persistent colorbar.
-        """
-        self._resolve_animation_clim(datasets, opts)
-        if colorbar:
-            self._animation_colorbar(opts, cbar_label)
-
-    def _draw_animation_frame(self, data: Any, kind: str, opts: dict, *, ocean: bool, coastlines: bool,
-                              title: Optional[str] = None) -> None:
-        """Draw one animation frame: optional ocean disc, the field, optional coastlines, optional title.
-
-        The per-frame body shared by :meth:`animate` and :meth:`rotate`. Ocean fill and coastlines are
-        decoration: the ocean disc is drawn only on a globe, and a coastline failure (no network/data) is
-        swallowed so the animation still renders.
-        """
-        if ocean and self.globe:
-            self.ocean()
-        getattr(self, kind)(data, **opts)
-        if coastlines:
-            try:
-                self.coastlines()
-            except Exception:  # network/data unavailable — decoration is best-effort
-                pass
-        if title is not None:
-            self.set_title(title)
-
-    def animate(self, stack: Any, *, kind: str = "imshow", fps: float = 3.0,
-                titles: Optional[Sequence[str]] = None, ocean: bool = False, coastlines: bool = False,
-                colorbar: bool = False, cbar_label: Optional[str] = None, **kwargs) -> FuncAnimation:
-        """Animate a stack of rasters over this map, returning a matplotlib :class:`FuncAnimation`.
-
-        Each frame reprojects ``stack[i]`` to the display CRS (pyramids), renders it with the ``kind`` method
-        (cleopatra), optionally draws the ocean disc / coastlines, and — on a globe — applies the projection
-        frame, so every frame gets the boundary, graticule, and limb-clipping for free. The returned
-        animation is lazy: call ``anim.save("out.gif", writer=PillowWriter(fps=...))`` or display it.
-
-        All frames share **one colour scale**: ``vmin``/``vmax`` from ``kwargs`` if given, else computed once
-        from the whole stack — so the colours (and the colorbar) do not flicker between frames.
-
-        Args:
-            stack: An ordered, indexable collection of pyramids ``Dataset`` frames (e.g. a list, or a
-                ``DatasetCollection`` datacube) — one raster per animation frame.
-            kind: The field method used to draw each frame — one of ``"imshow"`` / ``"contourf"`` /
-                ``"contour"`` / ``"pcolormesh"`` / ``"block"``.
-            fps: Frames per second (sets the inter-frame interval).
-            titles: Optional per-frame titles; must match the stack length when given.
-            ocean: When True, fill the ocean disc behind each frame (globe maps only).
-            coastlines: When True, overlay coastlines each frame (best-effort; ignored if unreachable).
-            colorbar: When True, add one static colorbar (drawn once, not per frame) using the shared
-                colour scale.
-            cbar_label: Optional label for the colorbar.
-            **kwargs: Forwarded to the ``kind`` method (e.g. ``cmap``, ``vmin``, ``vmax``).
-
-        Returns:
-            A :class:`matplotlib.animation.FuncAnimation` over ``len(stack)`` frames (also kept on
-            ``self._animation`` so it is not garbage-collected before you save/display it).
-
-        Raises:
-            ValueError: if ``kind`` is not a known field renderer, ``stack`` is empty, or ``titles`` is given
-                with a mismatched length.
-        """
-        if kind not in _ANIMATION_KINDS:
-            raise ValueError(f"unknown animation kind {kind!r}; choose one of {_ANIMATION_KINDS}")
-        frames = list(stack)
-        if not frames:
-            raise ValueError("animate got an empty stack (nothing to animate)")
-        if titles is not None and len(titles) != len(frames):
-            raise ValueError(f"titles length ({len(titles)}) must match the stack length ({len(frames)})")
-        self._prime_animation(frames, kwargs, colorbar=colorbar, cbar_label=cbar_label)
-
-        def draw_one(i: int) -> None:
-            title = titles[i] if titles is not None else None
-            self._draw_animation_frame(frames[i], kind, kwargs, ocean=ocean, coastlines=coastlines,
-                                       title=title)
-
-        return self._animate_frames(draw_one, len(frames), fps)
-
-    def rotate(self, dataset: Any, *, lat: float = 15.0, n_frames: int = 24, fps: float = 8.0,
-               lon0: float = -180.0, kind: str = "imshow", ocean: bool = False, coastlines: bool = False,
-               colorbar: bool = False, cbar_label: Optional[str] = None, **kwargs) -> FuncAnimation:
-        """Spin an orthographic globe over a single field by sweeping the centre longitude.
-
-        Forces a globe map and redraws ``dataset`` on ``n_frames`` orthographic projections whose centre
-        longitude steps a full 360 degrees from ``lon0``.
-
-        **Terminal for this Map's projection:** ``rotate`` sets ``globe=True`` and sweeps the display CRS
-        (:attr:`crs`) as the animation renders, leaving the Map centred on the **final** frame. Treat a
-        rotated Map as consumed by the animation — create a fresh ``Map`` if you need the original projection.
-
-        Args:
-            dataset: The pyramids ``Dataset`` to spin (reprojected per frame).
-            lat: Centre latitude of every orthographic view.
-            n_frames: Number of frames spanning the full 360-degree turn.
-            fps: Frames per second.
-            lon0: Starting centre longitude.
-            kind: The field method used to draw the data (``"imshow"`` / ``"contourf"`` / ``"pcolormesh"`` /
-                ``"contour"`` / ``"block"``).
-            ocean: When True, fill the ocean disc behind the data each frame.
-            coastlines: When True, overlay coastlines each frame (best-effort).
-            colorbar: When True, add one static colorbar (drawn once) using the shared colour scale.
-            cbar_label: Optional label for the colorbar.
-            **kwargs: Forwarded to the ``kind`` method (e.g. ``cmap``, ``vmin``, ``vmax``).
-
-        Returns:
-            A :class:`matplotlib.animation.FuncAnimation` over ``n_frames`` frames (also kept on
-            ``self._animation`` so it is not garbage-collected before you save/display it).
-
-        Raises:
-            ValueError: if ``n_frames`` is less than 1, or ``kind`` is not a known field renderer.
-        """
-        if n_frames < 1:
-            raise ValueError("rotate needs n_frames >= 1")
-        if kind not in _ANIMATION_KINDS:
-            raise ValueError(f"unknown animation kind {kind!r}; choose one of {_ANIMATION_KINDS}")
-        self.globe = True
-        self._prime_animation([dataset], kwargs, colorbar=colorbar, cbar_label=cbar_label)
-        lons = [lon0 + k * (360.0 / n_frames) for k in range(n_frames)]
-
-        def draw_one(i: int) -> None:
-            self.crs = projections.orthographic(lon=lons[i], lat=lat)
-            self._draw_animation_frame(dataset, kind, kwargs, ocean=ocean, coastlines=coastlines)
-
-        return self._animate_frames(draw_one, n_frames, fps)
